requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.2

# Vector Databases
faiss-cpu==1.7.4
chromadb==0.4.18
weaviate-client==3.25.3

# NLP and Embeddings
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
openai==1.3.7

# Text Processing
nltk==3.8.1
spacy==3.7.2
langchain==0.0.340
tiktoken==0.5.1

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0

# Visualization and UI
altair==5.1.2
pillow==10.1.0
wordcloud==1.9.2
networkx==3.2.1

# Data handling
openpyxl==3.1.2
xlsxwriter==3.1.9

